chore: move webhooks to match docs
